requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# NewsRecommendation\urls.py: 16,17
# NewsRecommendation\wsgi.py: 12
# algorithm\NaiveBayes\NaiveBayesPredict.py: 11
# algorithm\Recommendation\ContentBased.py: 11
# algorithm\Recommendation\UserCF.py: 12
# manage.py: 14
# news\admin.py: 1
# news\apps.py: 1
# news\migrations\0001_initial.py: 5
# news\migrations\0002_auto_20180131_1310.py: 5
# news\migrations\0003_auto_20180131_1513.py: 5
# news\migrations\0004_auto_20180201_1516.py: 6,7
# news\migrations\0005_auto_20180202_1313.py: 5
# news\migrations\0006_news_is_predict.py: 5
# news\migrations\0007_auto_20180207_1625.py: 5
# news\models.py: 4,5
# news\serializers.py: 9
# news\tests.py: 1
# news\urls.py: 14
# news\views.py: 7,14
# user\admin.py: 1,3
# user\apps.py: 1
# user\migrations\0001_initial.py: 6,7
# user\migrations\0002_user_search.py: 6,7
# user\migrations\0003_auto_20180207_1625.py: 6,7
# user\migrations\0004_auto_20180208_2320.py: 5
# user\models.py: 7,8
# user\serializers.py: 7
# user\tests.py: 1
# user\views.py: 1
Django == 1.11.5

#pymysql
pymysql

# spiders\ChinaNews\ChinaNews\items.py: 8
# spiders\ChinaNews\ChinaNews\middlewares.py: 8
# spiders\ChinaNews\ChinaNews\spiders\ChinaNews_spider.py: 4
# spiders\ChinaNews\start.py: 1
# spiders\toutiao\toutiao\items.py: 8
# spiders\toutiao\toutiao\middlewares.py: 8
Scrapy == 1.4.0

# spiders\ChinaNews\ChinaNews\pipelines.py: 8
Twisted == 17.5.0

# spiders\ChinaNews\ChinaNews\spiders\ChinaNews_spider.py: 9
beautifulsoup4 == 4.6.0

# news\serializers.py: 10
# news\urls.py: 7
# news\views.py: 8,10,15,16
# user\serializers.py: 8
djangorestframework == 3.7.3

# algorithm\NaiveBayes\NB.py: 10
# algorithm\NaiveBayes\NaiveBayesPredict.py: 10
# spiders\ChinaNews\ChinaNews\spiders\ChinaNews_spider.py: 10
jieba == 0.38

# algorithm\Recommendation\ContentBased.py: 10
numpy == 1.13.1

# algorithm\NaiveBayes\NB.py: 7,11
# algorithm\NaiveBayes\NaiveBayesPredict.py: 8
scikit_learn == 0.19.1
